requirements.txt
numpy>=1.14.2
pandas>=0.24.0
rasterio>=1.0.7
rasterstats>=0.13.0
shapely>=1.6.4
geopandas>=0.4.0
typer-cli>=0.0.12
PyYAML>=5.1
h3pandas>=0.2.2
